spark-backend/requirements-副本.txt
pwdlib[argon2]==0.2.1
fastapi-mail>=1.4.1
pydantic==2.12.5
pydantic-settings>=2.2.1
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
SQLAlchemy>=2.0.0
aiomysql>=0.2.0
aiosmtplib>=3.0.1
PyJWT>=2.8.0
langchain-deepseek>=0.1.4
langchain==1.1.0
langchain-community>=0.3.0
langchain-text-splitters>=0.3.0
chromadb>=0.4.0
pypdf>=3.0.0
docx2txt>=0.8
loguru>=0.7.0
python-dotenv>=1.0.0
pyyaml>=6.0.0
python-multipart>=0.0.9
dashscope>=1.14.0
redis>=5.0.0
httpx>=0.27.0